// src/content-cloud/schema.ts
export type JSON = any;
export type Timestamp = string;
// ...
export type BlockType = Entry &
	IndependentEntry & {
		fields: {
			dependencies?: JSON | null;
			langcode?: string | null;
			plugin?: string | null;
			provider?: string | null;
			region?: string | null;
			settings?: JSON | null;
			status?: boolean | null;
			theme?: string | null;
			thirdPartySettings?: JSON | null;
			uuid?: string | null;
			visibility?: JSON | null;
		};
	};
// ...
export type BlogContent = Entry &
	IndependentEntry & {
		fields: {
			author: Collection<PersonContent>;
			body?: TextWithSummaryProperty | null;
			/**
			 * Select up to two categories that best describe this content.
			 */
			category: CategoryTerm[];
			/**
			 * The time that the node was last edited.
			 */
			changed?: Timestamp | null;
			/**
			 * The date and time that the content was created.
			 */
			created?: Timestamp | null;
			introImage?: AnyEntry | null;
			langcode?: string | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			metaTags?: string | null;
			path?: PathProperty | null;
			promote?: boolean | null;
			/**
			 * Use this field to override the CTA text (50-character max) used when referencing this page. By default the text will be "Read More."
			 */
			referenceCtaText?: string | null;
			/**
			 * Fill out this field if the content could benefit from a lengthier description on the search results page and search engines.
			 */
			referenceDescription?: string | null;
			/**
			 * A brief description (160-character max) used to describe this content wherever it is referenced.
			 */
			referenceDescShort?: string | null;
			/**
			 * Upload an image that is at least 764px in width and 452px in height.
			 */
			referenceThumbnail?: AnyEntry | null;
			referenceTitle?: string | null;
			/**
			 * Briefly describe the changes you have made.
			 */
			revisionLog?: string | null;
			/**
			 * Specifies which action that Rabbit Hole should take.
			 */
			rhAction?: string | null;
			/**
			 * The path to where the user should get redirected to.
			 */
			rhRedirect?: string | null;
			/**
			 * Specifies the action that should be used when the redirect path is invalid or empty.
			 */
			rhRedirectFallbackAction?: string | null;
			/**
			 * Specifies the HTTP response code that should be used when perform a redirect.
			 */
			rhRedirectResponse?: number | null;
			/**
			 * Select up to two relevant roles/personas related to the content.
			 */
			role: RoleTerm[];
			status?: boolean | null;
			sticky?: boolean | null;
			title?: string | null;
			/**
			 * Select up to five topics that best describe this content.
			 */
			topics: TopicTerm[];
			type?: string | null;
			/**
			 * Select a vertical that best describes this content.
			 */
			vertical?: VerticalTerm | null;
		};
	};
// ...
export type BrandfolderImageMedia = Entry &
	IndependentEntry & {
		fields: {
			bfImage?: File | null;
			brandfolderAttachmentId?: string | null;
			bundle?: string | null;
			/**
			 * The time the media item was last edited.
			 */
			changed?: Timestamp | null;
			/**
			 * The time the media item was created.
			 */
			created?: Timestamp | null;
			imageBackground?: string | null;
			langcode?: string | null;
			/**
			 * Optionally link an image to another page. This is currently only used for logos.
			 */
			mediaImageLink?: LinkProperty | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			name?: string | null;
			path?: PathProperty | null;
			/**
			 * The time that the current revision was created.
			 */
			revisionCreated?: Timestamp | null;
			status?: boolean | null;
		};
	};
// ...
export type CategoryTerm = Entry &
	IndependentEntry &
	SystemTagContentEntry & {
		fields: {
			/**
			 * The time that the term was last edited.
			 */
			changed?: Timestamp | null;
			created?: Timestamp | null;
			description?: TextLongProperty | null;
			/**
			 * The term language code.
			 */
			langcode?: string | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			name?: string | null;
			parent?: CategoryTerm | null;
			path?: PathProperty | null;
			/**
			 * The time that the current revision was created.
			 */
			revisionCreated?: Timestamp | null;
			status?: boolean | null;
			vid?: string | null;
			/**
			 * The weight of this term in relation to other terms.
			 */
			weight?: number | null;
		};
	};
// ...
export type Collection<Type> = {
	limit: number;
	skip: number;
	total: number;
	items: Type[];
};
// ...
export type Entry = {
	assetEntry?: SystemAsset | null;
	externalLinks?: SystemExternalEntryLink[] | null;
	linkMetadata?: JSON | null;
	sys: Sys;
	tagEntry?: SystemTag | null;
};

export type EntryLinkInput = {
	contentTypeCustomId?: string | null;
	contentTypeId?: string | null;
	contentTypeMachineName?: string | null;
	customId?: string | null;
	id?: string | null;
	meta?: JSON | null;
	uuid?: string | null;
};

export type EntryMetadata = {
	tags?: SystemTag[] | null;
};
// ...
export type FlagsContentUserData = IndependentEntry & {
	fields: {
		bookmarked?: boolean | null;
		readAt?: Timestamp | null;
	};
};
// ...
export type IndependentEntry = Entry & {
	sys: {
		id: string;
		entryCreatedAt: string;
		entryVersion: number;
		isPublished: boolean;
		versionCreatedAt: string;
		versionId: string;
		type: 'Content';
	};
} & {
	flagsUserData?: FlagsContentUserData | null;
	metadata?: EntryMetadata | null;
};
// ...
export type LinkProperty = Entry & {
	fields: {
		entity?: AnyEntry | null;
		options?: JSON | null;
		title?: string | null;
		uri?: string | null;
	};
};
// ...
export type PathProperty = Entry & {
	fields: {
		alias?: string | null;
		langcode?: string | null;
	};
};

export type PersonContent = Entry &
	IndependentEntry & {
		fields: {
			body?: TextWithSummaryProperty | null;
			/**
			 * The time that the node was last edited.
			 */
			changed?: Timestamp | null;
			/**
			 * The date and time that the content was created.
			 */
			created?: Timestamp | null;
			langcode?: string | null;
			linkedinLink?: LinkProperty | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			path?: PathProperty | null;
			personCompany?: string | null;
			personPhoto?: AnyEntry | null;
			personPosition?: string | null;
			personType?: string[] | null;
			promote?: boolean | null;
			/**
			 * Briefly describe the changes you have made.
			 */
			revisionLog?: string | null;
			/**
			 * Specifies which action that Rabbit Hole should take.
			 */
			rhAction?: string | null;
			/**
			 * The path to where the user should get redirected to.
			 */
			rhRedirect?: string | null;
			/**
			 * Specifies the action that should be used when the redirect path is invalid or empty.
			 */
			rhRedirectFallbackAction?: string | null;
			/**
			 * Specifies the HTTP response code that should be used when perform a redirect.
			 */
			rhRedirectResponse?: number | null;
			status?: boolean | null;
			sticky?: boolean | null;
			title?: string | null;
			twitterLink?: LinkProperty | null;
			type?: string | null;
		};
	};
// ...
export type RoleTerm = Entry &
	IndependentEntry &
	SystemTagContentEntry & {
		fields: {
			/**
			 * The time that the term was last edited.
			 */
			changed?: Timestamp | null;
			created?: Timestamp | null;
			description?: TextLongProperty | null;
			/**
			 * The term language code.
			 */
			langcode?: string | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			name?: string | null;
			parent?: RoleTerm | null;
			path?: PathProperty | null;
			/**
			 * The time that the current revision was created.
			 */
			revisionCreated?: Timestamp | null;
			status?: boolean | null;
			vid?: string | null;
			/**
			 * The weight of this term in relation to other terms.
			 */
			weight?: number | null;
		};
	};
// ...
export type Sys = {
	contentType?: SystemLink | null;
	customId?: string | null;
	description?: string | null;
	entryCreatedAt?: Timestamp | null;
	entryVersion?: number | null;
	firstPublishedAt?: Timestamp | null;
	id?: string | null;
	isPublished?: boolean | null;
	locale?: SystemLink | null;
	localizationVersion?: number | null;
	name?: string | null;
	publishedAt?: Timestamp | null;
	publishedEntryVersion?: number | null;
	publishedLocalizationVersion?: number | null;
	revisionCreatedAt?: Timestamp | null;
	slug?: string | null;
	space: SystemLink;
	type: string;
	uuid?: string | null;
	versionId?: string | null;
};
// ...
export type SystemAsset = {
	customVersionId?: string | null;
	description?: string | null;
	downloadUrl?: string | null;
	embedUrl?: string | null;
	hash?: string | null;
	id: string;
	imageUrl?: string | null;
	mimeType?: string | null;
	mimeTypeGroup?: SystemMimeTypeGroup | null;
	name: string;
	size?: number | null;
	uuid?: string | null;
};
// ...
export type SystemExternalEntryLink = {
	authorEmail?: string | null;
	authorName?: string | null;
	canonicalUrl?: string | null;
	deleteUrl?: string | null;
	domain: string;
	editUrl?: string | null;
	environmentUuid?: string | null;
	externalTargetEntryId?: string | null;
	externalTargetVersionId?: string | null;
	linkType: SystemExternalEntryLinkType;
	prettyUrl?: string | null;
	privateMetadata?: JSON | null;
	publicMetadata?: JSON | null;
	publisherEmail?: string | null;
	publisherName?: string | null;
	siteUuid?: string | null;
	spaceUuid?: string | null;
	versionUrl?: string | null;
};
// ...
export type SystemTag = {
	description?: string | null;
	id: string;
	name: string;
	uuid?: string | null;
};

export type SystemTagContentEntry = {
	tagEntry?: SystemTag | null;
};

export type TextLongProperty = Entry & {
	fields: {
		embedded: Collection<AnyEntry>;
		format?: string | null;
		processed?: string | null;
		value?: string | null;
	};
};

export type TextWithSummaryProperty = Entry & {
	fields: {
		embedded: Collection<AnyEntry>;
		format?: string | null;
		processed?: string | null;
		summary?: string | null;
		value?: string | null;
	};
};
/**
 * Use topics to group related content into categories. Displayed within the Resource Center.
 */
export type TopicTerm = Entry &
	IndependentEntry &
	SystemTagContentEntry & {
		fields: {
			/**
			 * The time that the term was last edited.
			 */
			changed?: Timestamp | null;
			created?: Timestamp | null;
			description?: TextLongProperty | null;
			/**
			 * The term language code.
			 */
			langcode?: string | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			name?: string | null;
			parent?: TopicTerm | null;
			path?: PathProperty | null;
			/**
			 * The time that the current revision was created.
			 */
			revisionCreated?: Timestamp | null;
			status?: boolean | null;
			vid?: string | null;
			/**
			 * The weight of this term in relation to other terms.
			 */
			weight?: number | null;
		};
	};
// ...
export type VerticalTerm = Entry &
	IndependentEntry &
	SystemTagContentEntry & {
		fields: {
			/**
			 * The time that the term was last edited.
			 */
			changed?: Timestamp | null;
			created?: Timestamp | null;
			description?: TextLongProperty | null;
			/**
			 * The term language code.
			 */
			langcode?: string | null;
			/**
			 * The meta tags for the entity.
			 */
			metatag?: JSON | null;
			name?: string | null;
			parent?: VerticalTerm | null;
			path?: PathProperty | null;
			/**
			 * The time that the current revision was created.
			 */
			revisionCreated?: Timestamp | null;
			status?: boolean | null;
			vid?: string | null;
			/**
			 * The weight of this term in relation to other terms.
			 */
			weight?: number | null;
		};
	};
// ...
export type AnyEntry =
	| ArticleContent
	| BlockType
	| BlogContent
	| BrandfolderImageMedia
	| CaseStudyContent
	| CategoryTerm
	| CreditTypeTerm
	| DocumentContent
	| DocumentMedia
	| EbookContent
	| EventContent
	| File
	| ImageMedia
	| ImageTypeTerm
	| LandingPageContent
	| LinkProperty
	| PageContent
	| PathProperty
	| PersonContent
	| PressReleaseContent
	| ProductFeaturesTerm
	| RemoteVideoMedia
	| RoleTerm
	| SmartdateProperty
	| SolutionContent
	| TextLongProperty
	| TextWithSummaryProperty
	| TopicTerm
	| VerticalTerm
	| VideoMedia;